Add iteration to the Simulator and Engines.
Simulator can now perform multiple iterations of the engines. It does this with a new
context manager and iterator. The context manager creates a context in which StopIteration
exceptions are caught for terminating iteration. The iterator will cycle the Simulator.engines
until iteration is complete. Simulator takes a max_iteration keyword argument to define an upper
limit (default=1). Engine objects also have an attribute which is updated with the current iteration.